webDemo/requirements.txt
alabaster==0.7.12
anaconda-navigator==2.0.3

async-generator==1.10

backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1

bkcharts==0.2

boto==2.49.0
brotlipy==0.7.0
certifi==2021.5.30

clyent==1.2.2
comtypes==1.1.10
conda==4.10.3
conda-build==3.21.4

conda-verify==3.4.2

cycler==0.10.0
cytoolz==0.11.0

decorator==4.4.2

et-xmlfile==1.0.1

Flask==1.1.2

glob2==0.7
gmpy2==2.1.0b5

HeapDict==1.0.1

imagesize==1.2.0

ipython-genutils==0.2.0

jdcal==1.4.1

llvmlite==0.36.0
locket==0.2.0

mccabe==0.6.1
menuinst==1.4.16

mkl-fft==1.3.0
mkl-random==1.2.2
mkl-service==2.4.0

multipledispatch==0.6.0
navigator-updater==0.2.1

pandocfilters==1.4.2
parso==0.7.0

pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1


ply==3.11

pycurl==7.43.0.6

pyflakes==2.2.0

pyparsing==2.4.7
PyQt5==5.12.3
PyQt5_sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1

pytest==6.2.4
pytest-base-url==1.4.2
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-playwright==0.1.2
pytest-rerunfailures==10.1
pywin32==301

PyYAML==5.4.1

QtPy==1.9.0
requests-unixsocket==0.2.0

scikit-image==0.18.2

simplegeneric==0.8.1
sip==4.19.25
slugify==0.0.1

sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3

tables==3.6.1

unicodecsv==0.14.1

webencodings==0.5.1

wincertstore==0.2

xlwt==1.3.0
xmltodict==0.12.0
zict==2.0.0

